refactor: move the cache logic to a generic node/vite/shared/cache.ts
getCacheValue()/setCacheValue() read and merge JSON entries of
node_modules/.vike/cache.json, one top-level key per feature. logSkillHint.ts
only keeps its cache key.
